+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ![DIIVE](images/logo_diive1_256px.png)
 
+## v0.85.7 | XX Feb 2025
+
+### New features
+
+- `diive.pkgs.formats.meteo.FormatMeteoForFluxnetUpload`
+
+### Notebooks
+
+- Added new notebook `notebooks/Formats/FormatMeteoForEddyProFluxProcessing.ipynb`
+
 ## v0.85.6 | 25 Feb 2025
 
 ### New features